+++ PolarionMcpTools/GlobalUsings.cs
@@ -10,5 +10,3 @@ global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.AspNetCore.Routing;
 global using Microsoft.AspNetCore.Http;
 global using FluentResults;
-
-